Castor3D: Moved StylesHolder to its own file.
